Fix no dirs in stats and improve add_files in utils
* Early exit instead of crash when no dirs are passed to scripts/stats.sh
* allow to skip git-annex even when the number of files is < limit

+++ scripts/stats.sh
@@ -55,6 +55,11 @@ then
 	readarray -t _dirs
 fi
 
+if [[ ${#_dirs[@]} == 0 ]]
+then
+	exit 0
+fi
+
 rm -f files_count.stats
 rm -f disk_usage.stats
 find "${_dirs[@]}" -type d | sort -u | while read d
